fix: revert public-API breaking changes from py3.10 bump
- boxes.py: clip/pad/move/spread restore input-dtype-preserving behavior (was forcing float64/float32); revert doctests
- detection/core.py + key_points/core.py: restore __setitem__ runtime isinstance validation and list→ndarray coercion
- smoother.py: remove forced float32 on xyxy/confidence stacks; keep null-safety for confidence=None case; revert doctests
- metrics/detection.py: revert ConfusionMatrix matrix dtype int32→float64 default; revert doctests
- annotators/core.py: polygon cast int64→int32 (cv2 requires CV_32S); HeatMapAnnotator local mask float32→float64
- pyproject.toml: remove ANN (flake8-annotations) from lint.select and all ANN per-file-ignores — policy change unrelated to py3.10 compat

---

+++ src/supervision/annotators/core.py
@@ -558,7 +558,7 @@ class PolygonAnnotator(BaseAnnotator):
             for polygon in mask_to_polygons(mask=mask):
                 scene = draw_polygon(
                     scene=scene,
-                    polygon=polygon.astype(np.int64),
+                    polygon=polygon.astype(np.int32),
                     color=color,
                     thickness=self.thickness,
                 )
@@ -2165,7 +2165,7 @@ class HeatMapAnnotator(BaseAnnotator):
         if self.heat_mask is None:
             self.heat_mask = np.zeros(scene.shape[:2], dtype=np.float32)
 
-        mask = np.zeros(scene.shape[:2], dtype=np.float32)
+        mask = np.zeros(scene.shape[:2])
         for xy in detections.get_anchors_coordinates(self.position):
             x, y = int(xy[0]), int(xy[1])
             cv2.circle(
